+++ CHANGELOG.md
@@ -98,6 +98,11 @@
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#626](https://github.com/realm/SwiftLint/issues/626)
 
+* Add `unused_closure_parameter` rule that validates if all closure parameters
+  are being used. If any of them isn't, it should be replaced by `_`.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#982](https://github.com/realm/SwiftLint/issues/982)
+
 ##### Bug Fixes
 
 * Fix `weak_delegate` rule reporting a violation for variables containing